+++ client/src/contexts/PopupMessageContext.ts
@@ -4,4 +4,5 @@ export interface PopupMessageContextInterface {
   displayPopupMessage: (message: string, type: 'success' | 'error') => void;
 }
 
-export const PopupMessageContext = createContext<PopupMessageContextInterface | null>(null);
+const PopupMessageContext = createContext<PopupMessageContextInterface | null>(null);
+export default PopupMessageContext;