+++ client/src/hooks/usePopupMessage.ts
@@ -1,11 +1,11 @@
 import { useContext } from 'react';
-import { PopupMessageContext, PopupMessageContextInterface } from '../contexts/PopupMessageContext';
+import PopupMessageContext, { PopupMessageContextInterface } from '../contexts/PopupMessageContext';
 
 export default function usePopupMessage(): PopupMessageContextInterface {
   const context = useContext<PopupMessageContextInterface | null>(PopupMessageContext);
 
   if (!context) {
-    throw new Error('usePopupMessage must be used within a PopupMessageProvider');
+    throw new Error('usePopupMessage must be used within a PopupMessageProvider.');
   }
 
   return context;