+++ client/src/utils/errorUtils.ts
@@ -25,22 +25,10 @@ export function getAsyncErrorData(err: unknown): AsyncErrorData | null {
     return null;
   }
 
-  if (axiosError.status === 429) {
-    handleRateLimitReached();
-    return null;
-  }
-
   return {
     status: axiosError.status,
     errMessage: axiosError.response.data.message,
     errReason: axiosError.response.data.reason,
     errResData: axiosError.response.data.resData,
   };
-}
-
-function handleRateLimitReached(): void {
-  const { displayPopupMessage } = usePopupMessage();
-  displayPopupMessage('Too many requests.', 'error');
-
-  // TODO: implement further with a modal explaining things further to the user.
 }